active_bayes_repo/requirements.txt
numpy
matplotlib
torch
pyyaml
